reduce component size by removing duplicate code
