+++ packages/charts/react-charts/library/src/components/DeclarativeChart/PlotlyColorAdapter.ts
@@ -1,9 +1,10 @@
 import * as React from 'react';
 import { color as d3Color } from 'd3-color';
-import type { PieColors, Color, PlotData } from '@fluentui/chart-utilities';
+import type { PieColors, Color, PlotData, Layout } from '@fluentui/chart-utilities';
 import { isArrayOrTypedArray } from '@fluentui/chart-utilities';
 import { areArraysEqual } from '../../utilities/utilities';
 import { DataVizPalette, getColorFromToken, getNextColor } from '../../utilities/colors';
+import { scaleLinear as d3ScaleLinear } from 'd3-scale';
 
 type PlotlyColorway = 'plotly' | 'others';
 
@@ -147,4 +148,31 @@ export const getOpacity = (series: Partial<PlotData>, index: number): number => 
       ? (series.marker?.opacity as number[])[index % (series.marker?.opacity as number[]).length]
       : (series.marker?.opacity as number)
     : series.opacity ?? 1;
+};
+
+export const createColorScale = (
+  layout: Partial<Layout> | undefined,
+  series: Partial<PlotData>,
+  currentColorScale: ((value: number) => string) | undefined,
+) => {
+  if (
+    layout?.coloraxis?.colorscale?.length &&
+    isArrayOrTypedArray(series.marker?.color) &&
+    (series.marker?.color as Color[]).length > 0 &&
+    typeof (series.marker?.color as Color[])?.[0] === 'number'
+  ) {
+    const scale = layout?.coloraxis?.colorscale as Array<[number, string]>;
+    const colorValues = series.marker?.color as number[];
+    const [dMin, dMax] = [
+      layout?.coloraxis?.cmin ?? Math.min(...colorValues),
+      layout?.coloraxis?.cmax ?? Math.max(...colorValues),
+    ];
+
+    // Normalize colorscale domain to actual data domain
+    const scaleDomain = scale.map(([pos]) => dMin + pos * (dMax - dMin));
+    const scaleColors = scale.map(item => item[1]);
+
+    return d3ScaleLinear<string>().domain(scaleDomain).range(scaleColors);
+  }
+  return currentColorScale;
 };
